refactor: remove service worker exports
